tex_unit update

+++ driver/tests/tex_demo/common.h
@@ -9,7 +9,7 @@ struct kernel_arg_t {
   uint32_t src_height;
   uint32_t src_stride;
   uint32_t src_pitch;
-  uint32_t src_ptr;
+  uint8_t src_ptr;
   uint32_t dst_width;
   uint32_t dst_height;
   uint32_t dst_stride;
